Struct ruroonga_client::ResultParser

pub struct ResultParser {
    // some fields omitted
}

Methods

impl ResultParser

ResultParser

Groonga returns following array json:

success response: [[status, start_time, elapsed_time], [[[matched_columns], [Array([column1, type1], ...)], [Array([result1, result2, ...])]]]]

error response: [[status, start_time, elapsed_time, error_information, ...]]

fn new(json: String) -> ResultParser

fn get_raw_object(&mut self) -> Box<JFObject>

Get raw response result.

fn get_header(&mut self) -> JFObject

Return header elements in response.

Panics

Panics if response json is corrupted.

fn status(&mut self) -> Option<&i64>

Return status in response.

Panics

Panics if response json is corrupted.

fn start_time(&mut self) -> Option<&f64>

Return start time in response.

Panics

Panics if response json is corrupted.

fn elapsed_time(&mut self) -> Option<&f64>

Return elapsed time in response.

Panics

Panics if response json is corrupted.

fn matched_columns(&mut self) -> Option<i64>

Return a number of matched columns in response.

fn result(&mut self) -> Option<Vec<JFObject>>

Get result in response.

If request succeeded, it can get matched result array. Otherwise, one can get error messages.

Panics

Panics if response json is corrupted.

fn into_row(&mut self) -> Rows

Convert to Rows type and return its type values.